Removed everything related to django-profiles
django-profiles appears to be an obsolete app. Allegedly much of its
functionality has been included in django-registration, or was never
important. In any case, it can't be used, since it's obsolete.

However, model enhydris.hcore.models.UserProfile still remains. It's not
clear what its function is and whether it also should be removed, and
what kind of migration should be needed for existing installations.