Remove dbsync and redo migrations from scratch and
In order to remove dbsync, the migrations needed to be done from
scratch, because the hcore migrations had a dependency on the dbsync
app. The migration path for existing installations is to go to Enydris
version 0.3.0 (the previous commit), apply the migrations, then go to
Enhydris 0.3.1 (which should be this commit), empty the
django_migrations database table, migrate --fake, then go to Enhydris
0.4 and apply migrations. All this will be written more clearly in the
release notes.