Remove the is_active fields (fixes #8)
The Station and Instrument models both had an is_active field.
Apparently the original designers of the database thought that it would
be useful to make queries of, e.g., active stations, as opposed to all
stations (including obsolete ones).

However, the correctness of this field depends on the procedures each
organization has. Many organizations don't have a specific procedure for
obsoleting a station; a station merely falls out of use (e.g. an
overseer stops working and (s)he is never replaced). Therefore, it is
unlikely that someone will go and enter the correct value in the
is_active field. Even if an organization does have processes that could
ensure correctness of is_active, they could merely specify an end_date
to a station or instrument, and therefore is_active is superfluous.

Indeed, in all Hydroscope databases, the field seems to be randomly
chosen, and in openmeteo.org it makes even less sense, since it is an
open database whose users are expected to merely abandon their stations
and not care about "closing" them properly.

In this commit the is_active fields have been removed, but the
migration checks that they are not being used before going on to remove
them.