Fix error in geographical data of create_test_data()
In test_views.create_test_data() some test stations are created. Two
commits ago (451c2d2083) the "point" attribute for these test stations
was added; however, if "point" is specified, "srid" must also be not
null, otherwise the station detail view does not work properly (this was
discovered because of the unit test added in the previous commit).