Fix timezone confusion when downloading time series
When downloading time series, you can specify a start date and/or an end
date. These two were aware; they were getting the time zone of the time
series. However, the get_all_data() method returns a dataframe with a
naive index. The interpretation of the start and end dates was
incorrect. The download_timeseries view was fixed to account for this.