Remove old timeseries data view
There was an old timeseries data view that was being used by the
timeseries chart in the timeseries detail. It had several problems: 1)
The code was badly written and needed rewrite; 2) It was providing the
data even if the user did not have permission to download the time
series; 3) It was providing yet another way to get time series data when
there's already /api/tsdata.