Fix access to empty timeseries group detail page (fix #341)
When a user attempts to access the timeseries group detail page
that doesn't yet has a timeseries, an error page is displayed.
This is caused by a code in the `chartsjs.html` template that
attempts to access the primary key of a default timeseries (which
does not exist yet) for the base url.

This commit resolves this bug by checking the existance of any related
timeseries before loading the `chartjs.html` template.