Fix HTML id mismatch in station detail page tabs (fixes #347)
The tab link for the `Data` tab menu was pointing to a tab content
with an expected HTML id of `timeseries`. However, the tab content
actual id was `timeseries_groups`. This id mismatch was the source
of the bug that resulted on the tab menu link not opening its
corresponding tab content. Thus, change the tab menu link to point
to `timeseries_groups` HTML id so it can successfully target the
right tab content.