Remove allauth and registration through API
We had implemented registration through API, but this is causing
problems. The most important is that it uses django-allauth, which
notably has many-to-one email addresses, which causes headaches. It adds
redundancy, because it keeps these email addresses in a separate table,
with the option of marking one of them as primary, and ignores the fact
that django.contrib.auth.models.User also has an email field. If a user
is created by an non-allauth-aware app (such as the Django admin, or
django-registration-redux), then at some point we get a "missing email"
exception, because something expects to find the user's email address in
allauth's table.

Since we will implement registration with django-registration-redux, and
to avoid such headaches, we remove registration through API since it
isn't useful at this point, and we also remove everything related to
allauth. However there's some dj_rest_auth functionality (login/logout
and more) that we don't remove. Some of it may be used, and some may be
useful for the future (and is not conflicting).

Should we ever need to bring registration through API back, most of the
work is the stuff removed by this commit.

In addition, should we ever need to use dj_rest_auth.registration
(for registration through API) and django-registration-redux (for
registration through templates) together, this settings.py snippet may
come in handy:

    # We use both dj_rest_auth.registration and django-registration-redux. These apps both
    # have the label "registration", resulting in conflict. We programmatically create a
    # dj_rest_auth.registration.apps module containing an AppConfig to resolve the problem.
    spec = importlib.util.spec_from_loader("dj_rest_auth.registration.apps", loader=None)
    module = importlib.util.module_from_spec(spec)
    code = textwrap.dedent(
        """\
        from django.apps import AppConfig
        class RestAuthRegistrationAppConfig(AppConfig):
            name = "dj_rest_auth.registration"
            label = "dj_rest_auth_registration"
        """
    )
    exec(code, module.__dict__)
    sys.modules["dj_rest_auth.registration.apps"] = module