Fix navbar covering part of the page
enhydris.css, in the page header (i.e. the navbar) had
"position: absolute" and was partially covering the page content. This
was because all pages had a map, and we do want the map to be partially
covered. However there are pages (such as registration/login/etc, about
to be added, or pages by add-on apps, like enhydris-synoptic's station
report) that shouldn't be partially covered and need
"position: relative".

The solution is to change enhydris.css to "position: relative" and add
"position: absolute" to the main_with_map template, since all pages with
a map inherit from that one.