Do not read all data from db if a slice is requested
The Timeseries.get_data() method accepts start_date and end_date.
However, this would slice the cached time series; if the cache was
empty, it would always read the entire data from the database into
the cache.

This worked well for cases like viewing time series on the web, because
in order to show the chart, it needed to retrieve the entire time series
anyway; subsequently, when playing with the chart in order to show
different sections, the entire time series was already in the cache.

However, it was failing in other cases, notable on enhydris-synoptic,
which needs to read the last 24 hours of many time series. Since
enhydris-synoptic runs, say, every 10 minutes, and many (or most) of the
time series it needs to read are also updated every 10 minutes, each
time it was running the cache was invalidated. So although it needed to
access only something like the last 144 rows of a time series, the
entire time series needed to be read from the database into the cache.
Several time series had a length of roughly a million rows and were
taking half a minute or more, when all that was needed was the last 144
rows. For all synoptic reports together, it would need something like 45
minutes to run (thus blocking the celery queue which grew and grew and
grew).

We fixed this problem by implementing a smarter caching system. Now it
only retrieves (and caches) the part of the time series requested. In
subsequent requests for data, if the part between the specified
start_date and end_date are already in the cache, it returns an
appropriate slice from the cache; otherwise it re-retrieves an
appropriate slice of the data from the database into the cache.