Improve authorization to access to timeseries data (fixes #487)
When a timeseries group is marked as `public`, it is always
available for viewing and downloading the timeseries data,
overwriting whatever the current value of the `ENHYDRIS_OPEN_CONTENT`
setting value. For logged-in users, the timeseries data is only
available for viewing and downloading if they are either the owner
or one of the maintainers of the station.