Allow many aggregated time series with the same step (fixes DEV-48)
Sometimes, in a given time series group, we may need to create many time
series with the same type and time step. For example, for temperature,
we may need the mean monthly temperature and the max monthly
temperature. Until now this was not possible because it would
violate the primary key constraint for the time series,
(timeseries_group, type, time_step). To solve this problem, the "name"
attribute was introduced and was made part of the primary key.