Reduce cache race conditions for time series data (fixes DEV-22)
In some rare cases, there was an internal server error when trying to
read a time series, because it would, for example, read the (cached)
start date and one zillionth of a second later it would read the
(cached) end date, and the latter was null although the first one was
not null, something that will never happen unless the data changes
between the two statements.

In order to make this better, instead of caching separately the start
date, end date, and time series records, we cache a single hash object
that contains start_date, end_date and time series data. The caching
part has also been separated into a different class, which is an
improvement in the code structure.

However, I've had second thoughts. The problem is that now a large part
of data is read many times from the cache, because every time we need
"start_date" we also get the cached time series records, which can be
tens of MB, and we can be reading start_date and end_date many times.

Therefore I'll not merge this.