Fix conflicting development dependency (fixes DEV-50)
The latest version of django-cors-headers in requirements-dev.txt had a
requirement on a recent version of Django and conflicted with the lower
version that was specified on requirements.txt. This was resulting in
uninstalling the correct version of Django and installing a more recent
one when "pip install -r requirements-dev.txt" was running. Fixed by
constraining the version of django-cors-headers.