When aggregating, recalculate last record if partial (fixes DEV-62)
Until now, when aggregating, it ignored and did not store the last
record of the aggregated time series if it contained the MISS flag.
This was because more source time series data might become available
later. However it also meant that if more source time series data never
became available, the last aggregated record would never be calculated.
In addition, we sometimes want to get an estimation of the current daily
value a bit before midnight.

Now the last record is calculated alright given enough data (i.e. if
the missing data are up to max_missing), and it is recalculated when
more source data become available.

This commit also upgrades pthelma to 2.1, and, correspondingly, changes
the missing flag from "MISS" to "MISSING3", where, in this example, 3 is
the number of missing values in the source data. This will not migrate
existing records. This lack of migration should not cause any problems
with aggregation.